+++ locales/es.json
@@ -874,6 +874,10 @@
         "user_error_message": "No se pudo inicializar el pago. Inténtalo de nuevo más tarde o usa otro método de pago",
         "payment_unavailable": "El método de pago no está disponible. Inténtalo de nuevo más tarde."
       },
+      "datatrans_lightbox": {
+        "title": "Pago seguro",
+        "description": "Será redirigido a una ventana de pago segura de Datatrans para completar su compra."
+      },
       "authorize_net": {
         "user_agreement_link": "Acuerdo de usuario",
         "privacy_policy_link": "Política de privacidad",
