+++ locales/fi.json
@@ -874,6 +874,10 @@
         "user_error_message": "Maksua ei voi aloittaa. Yritä uudelleen myöhemmin tai käytä toista maksutapaa",
         "payment_unavailable": "Maksutapa ei ole saatavilla. Yritä uudelleen myöhemmin"
       },
+      "datatrans_lightbox": {
+        "title": "Turvallinen maksu",
+        "description": "Sinut ohjataan turvalliseen Datatrans-maksuikkunaan ostoksesi viimeistelemiseksi."
+      },
       "authorize_net": {
         "user_agreement_link": "Käyttäjäsopimus",
         "privacy_policy_link": "Tietosuojakäytäntö",
