+++ locales/it.json
@@ -874,6 +874,10 @@
         "user_error_message": "Impossibile inizializzare il pagamento. Riprova più tardi o usa un altro metodo di pagamento",
         "payment_unavailable": "Metodo di pagamento non disponibile. Riprova più tardi."
       },
+      "datatrans_lightbox": {
+        "title": "Pagamento sicuro",
+        "description": "Verrai reindirizzato a una finestra di pagamento Datatrans sicura per completare il tuo acquisto."
+      },
       "authorize_net": {
         "user_agreement_link": "Accordo utente",
         "privacy_policy_link": "Informativa sulla privacy",
