+++ locales/pl.json
@@ -874,6 +874,10 @@
         "user_error_message": "Nie można zainicjować płatności. Spróbuj ponownie później lub użyj innej metody płatności",
         "payment_unavailable": "Metoda płatności jest niedostępna. Spróbuj ponownie później."
       },
+      "datatrans_lightbox": {
+        "title": "Bezpieczna płatność",
+        "description": "Zostaniesz przekierowany do bezpiecznego okna płatności Datatrans, aby dokończyć zakup."
+      },
       "authorize_net": {
         "user_agreement_link": "Umowa użytkownika",
         "privacy_policy_link": "Polityka prywatności",
