+++ locales/ru.json
@@ -874,6 +874,10 @@
         "user_error_message": "Не удалось инициализировать платёж. Попробуйте позже или выберите другой способ оплаты",
         "payment_unavailable": "Способ оплаты недоступен. Повторите попытку позже."
       },
+      "datatrans_lightbox": {
+        "title": "Безопасная оплата",
+        "description": "Вы будете перенаправлены в безопасное окно оплаты Datatrans для завершения покупки."
+      },
       "authorize_net": {
         "user_agreement_link": "Пользовательского соглашения",
         "privacy_policy_link": "Политике конфиденциальности",
