+++ locales/zh.json
@@ -874,6 +874,10 @@
         "user_error_message": "无法初始化支付。请稍后再试或使用其他支付方式",
         "payment_unavailable": "支付方式不可用，请稍后重试"
       },
+      "datatrans_lightbox": {
+        "title": "安全支付",
+        "description": "您将被重定向到安全的Datatrans支付窗口以完成购买。"
+      },
       "authorize_net": {
         "user_agreement_link": "用户协议",
         "privacy_policy_link": "隐私政策",
